Put Flame graph and Node graph related files into separate folders
